Move array length extraction for GenC compilation before ghost elimination phase
