+++ core/src/main/scala/stainless/genc/GenerateC.scala
@@ -9,11 +9,11 @@ import phases._
 
 object GenerateC {
 
-  def pipeline(implicit ctx: inox.Context) =
+  def pipeline(arrayLengthsMap: Map[Identifier, Int])(implicit ctx: inox.Context) =
     NamedLeonPhase("GhostElimination", GhostEliminationPhase(ctx)) andThen
     NamedLeonPhase("TrimSymbols", TrimSymbols(ctx)) andThen
     NamedLeonPhase("ComputeFunCtx", LeonPipeline.both(NoopPhase[Symbols], ComputeFunCtxPhase(ctx))) andThen
-    NamedLeonPhase("Scala2IR", Scala2IRPhase(ctx)) andThen
+    NamedLeonPhase("Scala2IR", Scala2IRPhase(arrayLengthsMap)) andThen
     NamedLeonPhase("Normalisation", NormalisationPhase(ctx)) andThen
     NamedLeonPhase("Lifting", LiftingPhase(ctx)) andThen
     NamedLeonPhase("Referencing", ReferencingPhase(ctx)) andThen
@@ -21,6 +21,10 @@ object GenerateC {
     NamedLeonPhase("IR2C", IR2CPhase(ctx)) andThen
     CFileOutputPhase(ctx)
 
-  def run(symbols: Symbols)(implicit ctx: inox.Context) = pipeline.run(symbols)
+  def run(symbols: Symbols)(implicit ctx: inox.Context) = {
+    // extract lengths before `GhostElimination` erases them
+    val arrayLengthsMap = ArraysLengthsExtraction.get(symbols)
+    pipeline(arrayLengthsMap).run(symbols)
+  }
 
 }
